prov/verbs: Move SRQ code to EP/MSG code
The SRQ code was moved from verbs_srq.c file to verbs_msg_ep.c file.
This is needed for futher separation of common code into common
functions and re-use this common stuff.
